[Refactor/#23]: Auth 폴더 분리
