--- src/features/Auth/index.ts
@@ -1,25 +1,0 @@
-import GenericForm from './components/common/Form/GenericForm';
-import FormTitle from './components/common/FormTitle/FormTitle';
-import TitleAndStep from './components/common/FormTitle/TitleAndStep';
-import FindPasswordForm from './components/FindPassword/FindPasswordForm';
-import LoginForm from './components/Login/LoginForm';
-import JoinForm from './components/Join/JoinForm';
-import { useFunnel } from './hooks/useFunnel';
-import { LoginForm as LoginFormType } from './types/login.d';
-import { useLogin } from './hooks/useLogin';
-import { useFindPassword } from './hooks/useFindPassword';
-import { FindPasswordFormType } from './types/findPassword.d';
-
-export {
-  LoginForm,
-  FindPasswordForm,
-  GenericForm,
-  FormTitle,
-  TitleAndStep,
-  useFunnel,
-  JoinForm,
-  useLogin,
-  useFindPassword,
-};
-
-export type { LoginFormType, FindPasswordFormType };
