+++ src/features/FindPassword/components/AuthWithIdForm.tsx
@@ -1,8 +1,8 @@
+import AuthForm from '@/components/Form/AuthForm';
 import FormInput from '@/components/Input/FormInput';
+import { REGEXP } from '@/constants/form/form';
 import { useFormContext } from 'react-hook-form';
 import styled from 'styled-components';
-import AuthForm from '../common/Form/AuthForm';
-import { REGEXP } from '../../constants/form';
 
 function AuthWithIdForm() {
   const { register } = useFormContext();
