+++ src/features/FindPassword/index.ts
@@ -1,0 +1,3 @@
+export { useFindPassword } from './hooks/useFindPassword';
+export { default as FindPasswordForm } from './components/FindPasswordForm';
+export type { FindPasswordFormType } from './types/findPassword.d';
