+++ src/features/Join/index.ts
@@ -1,0 +1,2 @@
+export { default as JoinForm } from './components/JoinForm';
+export { signup } from './api/signup.api';
