+++ src/types/funnel.d.ts
